git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/HEAD/root@7349 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ source/java/org/alfresco/web/bean/wcm/ImportWebsiteDialog.java
@@ -310,7 +310,10 @@ public class ImportWebsiteDialog
          }
          finally
          {
-            ImporterActionExecuter.deleteDir(tempDir);
+            if (tempDir.exists())
+            {
+               ImporterActionExecuter.deleteDir(tempDir);
+            }
          }
       }
       catch (IOException e)
@@ -330,6 +333,7 @@ public class ImportWebsiteDialog
       ServiceRegistry services = Repository.getServiceRegistry(FacesContext.getCurrentInstance());
       MimetypeService mimetypeService = services.getMimetypeService();
       File topdir = new File(dir);
+      if (!topdir.exists()) return;
       for (File file : topdir.listFiles())
       {
          try
